requirements.txt
alembic==1.0.11
Click==7.0
Flask==1.1.0
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.13
MarkupSafe==1.1.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.5
Werkzeug==0.15.4
